Simplify a little bit
getByte: b = bytes[i]; i++
The panic is in the runtime, already, and you don't handle it.

Encode: why does it receive a *[]byte, if it does not modify it?

Is the check of the next 2 bytes with isAsii in Encode OK?
You check that there is +1, +2 bytes remaining, but the same bytes' isAscii...